+++ src/features/reader/viewer/ReaderViewer.tsx
@@ -51,6 +51,7 @@ import {
     useReaderSettingsStore,
 } from '@/features/reader/stores/ReaderStore.ts';
 import { STABLE_EMPTY_OBJECT } from '@/base/Base.constants.ts';
+import { getPage } from '@/features/reader/overlay/progress-bar/ReaderProgressBar.utils.tsx';
 
 const READING_MODE_TO_IN_VIEWPORT_TYPE: Record<ReadingMode, PageInViewportType> = {
     [ReadingMode.SINGLE_PAGE]: PageInViewportType.X,
@@ -130,11 +131,11 @@ const BaseReaderViewer = ({
     }, [scrollbarXSize, scrollbarYSize]);
 
     const imageRefs = useRef<(HTMLElement | null)[]>(pages.map(() => null));
-    const [{ minChapterViewWidth, minChapterViewHeight, minChapterSizeSourceChapterId }, setChapterViewerSize] =
+    const [{ minChapterViewerWidth, minChapterViewerHeight, minChapterViewerSourceChapterId }, setChapterViewerSize] =
         useState({
-            minChapterViewWidth: 0,
-            minChapterViewHeight: 0,
-            minChapterSizeSourceChapterId: -1,
+            minChapterViewerWidth: 0,
+            minChapterViewerHeight: 0,
+            minChapterViewerSourceChapterId: -1,
         });
 
     const [, setTriggerReRender] = useState({});
@@ -158,41 +159,41 @@ const BaseReaderViewer = ({
         [currentChapter, chaptersToRender],
     );
 
-    const onChapterViewSizeChange = useCallback(
+    const onChapterViewerSizeChange = useCallback(
         (width: number, height: number, chapterId: ChapterIdInfo['id']) => {
             if (!isContinuousReadingModeActive) {
                 return;
             }
 
-            const isSameChapterId = chapterId === minChapterSizeSourceChapterId;
+            const isSameChapterId = chapterId === minChapterViewerSourceChapterId;
 
             if (isContinuousVerticalReadingModeActive) {
-                if (!isSameChapterId && minChapterViewWidth >= width) {
+                if (!isSameChapterId && minChapterViewerWidth >= width) {
                     return;
                 }
 
                 setChapterViewerSize({
-                    minChapterViewWidth: width,
-                    minChapterViewHeight: 0,
-                    minChapterSizeSourceChapterId: chapterId,
+                    minChapterViewerWidth: width,
+                    minChapterViewerHeight: 0,
+                    minChapterViewerSourceChapterId: chapterId,
                 });
                 return;
             }
 
-            if (isSameChapterId || minChapterViewHeight < height) {
+            if (isSameChapterId || minChapterViewerHeight < height) {
                 setChapterViewerSize({
-                    minChapterViewWidth: 0,
-                    minChapterViewHeight: height,
-                    minChapterSizeSourceChapterId: chapterId,
+                    minChapterViewerWidth: 0,
+                    minChapterViewerHeight: height,
+                    minChapterViewerSourceChapterId: chapterId,
                 });
             }
         },
         [
             isContinuousReadingModeActive,
             isContinuousVerticalReadingModeActive,
-            minChapterViewWidth,
-            minChapterViewHeight,
-            minChapterSizeSourceChapterId,
+            minChapterViewerWidth,
+            minChapterViewerHeight,
+            minChapterViewerSourceChapterId,
         ],
     );
 
@@ -233,9 +234,9 @@ const BaseReaderViewer = ({
 
     useLayoutEffect(() => {
         setChapterViewerSize({
-            minChapterViewWidth: 0,
-            minChapterViewHeight: 0,
-            minChapterSizeSourceChapterId: -1,
+            minChapterViewerWidth: 0,
+            minChapterViewerHeight: 0,
+            minChapterViewerSourceChapterId: -1,
         });
         setTriggerReRender({});
     }, [readingMode]);
@@ -286,14 +287,19 @@ const BaseReaderViewer = ({
                 const nextChapter =
                     chaptersToRender[chapterIndex - 1] ?? chapters[initialChapterIndex - visibleChapters.trailing - 1];
 
+                const isChapterViewerSourceChapter = chapter.id === minChapterViewerSourceChapterId;
+
                 const isInitialChapter = chapter.id === initialChapter.id;
                 const isCurrentChapter = chapter.id === currentChapter.id;
                 const isPreviousChapter = chapter.id === chaptersToRender[currentChapterIndex + 1]?.id;
                 const isNextChapter = chapter.id === chaptersToRender[currentChapterIndex - 1]?.id;
+                const isAdjacentChapterToCurrentChapter = isPreviousChapter || isNextChapter;
+
                 const isLeadingChapter = initialChapter.sourceOrder > chapter.sourceOrder;
                 const isTrailingChapter = initialChapter.sourceOrder < chapter.sourceOrder;
                 const isLastLeadingChapter = visibleChapters.lastLeadingChapterSourceOrder === chapter.sourceOrder;
                 const isLastTrailingChapter = visibleChapters.lastTrailingChapterSourceOrder === chapter.sourceOrder;
+
                 const isPreloadMode =
                     (isLastLeadingChapter && visibleChapters.isLeadingChapterPreloadMode) ||
                     (isLastTrailingChapter && visibleChapters.isTrailingChapterPreloadMode);
@@ -304,7 +310,28 @@ const BaseReaderViewer = ({
                     visibleChapters,
                 );
 
-                const isChapterSizeSourceChapter = chapter.id === minChapterSizeSourceChapterId;
+                const currentChapterCurrentPageIndex = isAdjacentChapterToCurrentChapter
+                    ? getReaderChapterViewerCurrentPageIndex(
+                          currentPageIndex,
+                          currentChapter,
+                          currentChapter,
+                          true,
+                          isCurrentChapterReady,
+                          initialChapter.sourceOrder > currentChapter.sourceOrder,
+                          initialChapter.sourceOrder < currentChapter.sourceOrder,
+                          visibleChapters,
+                      )
+                    : -1;
+                const currentChapterCurrentPagesIndex = isAdjacentChapterToCurrentChapter
+                    ? getPage(currentChapterCurrentPageIndex, pages).pagesIndex
+                    : -1;
+
+                const currentChapterRemainingLeadingPages = currentChapterCurrentPagesIndex;
+                const currentChapterRemainingTrailingPages = pages.length - 1 - currentChapterCurrentPagesIndex;
+
+                const currentChapterRemainingPages = isNextChapter
+                    ? currentChapterRemainingTrailingPages
+                    : currentChapterRemainingLeadingPages;
 
                 return (
                     <ReaderChapterViewer
@@ -315,6 +342,7 @@ const BaseReaderViewer = ({
                         isPreviousChapterVisible={previousNextChapterVisibility.previous}
                         isNextChapterVisible={previousNextChapterVisibility.next}
                         lastPageRead={coerceIn(chapter.lastPageRead, 0, chapter.pageCount - 1)}
+                        currentChapterRemainingPages={currentChapterRemainingPages}
                         currentPageIndex={getReaderChapterViewerCurrentPageIndex(
                             currentPageIndex,
                             chapter,
@@ -356,9 +384,9 @@ const BaseReaderViewer = ({
                         customFilter={customFilter}
                         shouldStretchPage={shouldStretchPage}
                         readerNavBarWidth={readerNavBarWidth}
-                        onSizeChange={onChapterViewSizeChange}
-                        minWidth={isChapterSizeSourceChapter ? 0 : minChapterViewWidth}
-                        minHeight={isChapterSizeSourceChapter ? 0 : minChapterViewHeight}
+                        onSizeChange={onChapterViewerSizeChange}
+                        minWidth={isChapterViewerSourceChapter ? 0 : minChapterViewerWidth}
+                        minHeight={isChapterViewerSourceChapter ? 0 : minChapterViewerHeight}
                         scrollElement={scrollElementRef.current}
                     />
                 );